新增产品列表

+++ src/main/java/com/wwt/managemail/vo/BankBillQuery.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class BankBillQuery {
-    private Long bankCardId;
+    private Integer bankCardId;
     private String transactionType;
 }
